Replace rug/GMP with num-bigint
This builds on some of the work done by @codec-abc, with a bunch of
changes to make sure the various shifting operations work as before.

While num-bigint may perform worse compared to GMP, it has a few
benefits:

1. GMP/rug is having trouble compiling on Windows, due to a Rust bug:
   https://github.com/rust-lang/rust/issues/47048.

2. When using GMP you can no longer cross compile the VM, or at least
   not as easy as `cargo build --target=X`. This makes it more difficult
   to for example build a musl build without having to use Docker.

3. GMP takes a while to compile, much longer than num-bigint. This isn't
   _that_ big of a deal, but combined with Rust's already slow compile
   times I'd prefer to keep compilation as fast as possible.

4. num-bigint is pure Rust. In _theory_ this should make it easier to
   distribute, as we don't have to worry about also compiling C code. In
   practise I think this won't matter that much since GMP is such a
   mature library.

+++ vm/src/hasher.rs
@@ -1,7 +1,7 @@
 //! Types and methods for hashing objects.
 #![cfg_attr(feature = "cargo-clippy", allow(new_without_default_derive))]
 
-use rug::Integer;
+use num_bigint::BigInt;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher as HasherTrait};
 use std::i64;
@@ -36,7 +36,7 @@ impl Hasher {
         self.write_integer(bits);
     }
 
-    pub fn write_bigint(&mut self, value: &Integer) {
+    pub fn write_bigint(&mut self, value: &BigInt) {
         value.hash(&mut self.hasher);
     }
