+++ vm/src/vm/state.rs
@@ -4,8 +4,8 @@
 //! garbage collections, the configuration, the files that have been parsed,
 //! etc.
 
+use num_bigint::BigInt;
 use parking_lot::Mutex;
-use rug::Integer;
 use std::sync::{Arc, RwLock};
 use std::time;
 
@@ -235,7 +235,7 @@ impl State {
         alloc.allocate_with_prototype(value, self.integer_prototype)
     }
 
-    pub fn allocate_permanent_bigint(&self, bigint: Integer) -> ObjectPointer {
+    pub fn allocate_permanent_bigint(&self, bigint: BigInt) -> ObjectPointer {
         let mut alloc = self.permanent_allocator.lock();
         let value = object_value::bigint(bigint);
